fix: clippy warnings surfaced by CI
Four clippy lints broke the dev branch push:
- sp-core/src/lib.rs: const_is_empty on CARGO_PKG_VERSION (always true,
  tautological assert removed)
- sp-decoder/audio/symphonia_reader.rs: unnecessary_cast (t.seconds is
  already u64 in symphonia 0.5)
- sp-decoder/split_sync.rs: while_let_loop rewrite of the audio drain
- sp-server/db/models.rs: too_many_arguments on mark_video_processed_pair
  (eight-arg SQL helper, #[allow] added)
- sp-server/downloader/normalize.rs: vec_init_then_push rewritten to
  vec![] macro form
- sp-server/api/routes_tests.rs: needless_borrows_for_generic_args on
  format!() argument to .uri()

+++ crates/sp-decoder/src/audio/symphonia_reader.rs
@@ -79,7 +79,7 @@ impl SymphoniaAudioReader {
         let duration_ms = match (codec_params.n_frames, codec_params.time_base) {
             (Some(n_frames), Some(tb)) => {
                 let t = tb.calc_time(n_frames);
-                (t.seconds as u64) * 1_000 + ((t.frac * 1_000.0) as u64)
+                t.seconds * 1_000 + ((t.frac * 1_000.0) as u64)
             }
             _ => 0,
         };
